BackEnd/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
environ==1.0
idna==3.4
inflection==0.5.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
xgboost==2.0.2
